+++ assets/javascripts/kitten/karl/grid/side-grid.js
@@ -8,19 +8,19 @@ export const KarlSideGridWithRightAside = () => {
     <SideGrid>
       <SideGridContent>
         <p>
-          Cat ipsum dolor sit amet, lick sellotape and meowzer!
-          Hopped up on catnip. Purr kitty power! and damn that dog or poop on
-          grasses so rub whiskers on bare skin act innocent meowing non stop
-          for food thug cat.
+          Has autem provincias, quas Orontes ambiens amnis imosque pedes Cassii
+          montis illius celsi praetermeans funditur in Parthenium mare, Gnaeus
+          Pompeius superato Tigrane regnis Armeniorum abstractas dicioni Romanae
+          coniunxit.
         </p>
       </SideGridContent>
 
       <SideGridAside>
         <p>
-          Cat ipsum dolor sit amet, lick sellotape and meowzer!
-          Hopped up on catnip. Purr kitty power! and damn that dog or poop on
-          grasses so rub whiskers on bare skin act innocent meowing non stop
-          for food thug cat.
+          Et eodem impetu Domitianum praecipitem per scalas itidem funibus constrinxerunt,
+          eosque coniunctos per ampla spatia civitatis acri raptavere discursu.
+          iamque artuum et membrorum divulsa conpage superscandentes corpora
+          mortuorum ad ultimam truncata deformitatem velut exsaturati.
         </p>
       </SideGridAside>
     </SideGrid>
@@ -32,19 +32,17 @@ export const KarlSideGridWithLeftAsideAndSmallSize = () => {
     <SideGrid asidePosition="start" asideSize="small">
       <SideGridAside>
         <p>
-          Cat ipsum dolor sit amet, lick sellotape and meowzer!
-          Hopped up on catnip. Purr kitty power! and damn that dog or poop on
-          grasses so rub whiskers on bare skin act innocent meowing non stop
-          for food thug cat.
+          Nec vox accusatoris ulla licet subditicii in his malorum quaerebatur
+          acervis ut saltem specie tenus crimina praescriptis legum committerentur,
+          quod aliquotiens fecere principes saevi: sed quicquid Caesaris.
         </p>
       </SideGridAside>
 
       <SideGridContent>
         <p>
-          Cat ipsum dolor sit amet, lick sellotape and meowzer!
-          Hopped up on catnip. Purr kitty power! and damn that dog or poop on
-          grasses so rub whiskers on bare skin act innocent meowing non stop
-          for food thug cat.
+          Haec igitur prima lex amicitiae sanciatur, ut ab amicis honesta petamus,
+          amicorum causa honesta faciamus, ne exspectemus quidem, dum rogemur;
+          studium semper adsit, cunctatio absit; consilium vero dare audeamus libere.
         </p>
       </SideGridContent>
     </SideGrid>